add schema parameter
add schema param for student selection, targets, model prep, inf prep, and checkpoints
