+++ pipelines/pdp/resources/github_pdp_inference.yml
@@ -120,7 +120,9 @@ resources:
               - --job_type
               - "{{job.parameters.job_type}}"
               - --config_file_path
-              - "{{tasks.data_ingestion.values.config_file_path}}"              
+              - "{{tasks.data_ingestion.values.config_file_path}}"
+              - --inst_schema
+              - "{{job.parameters.inst_schema}}"
           job_cluster_key: edvise-pdp-inference-pipeline-cluster
           libraries:
             - pypi:
@@ -194,6 +196,8 @@ resources:
               - "{{job.parameters.term_filter}}"
               - --config_file_path
               - "{{tasks.data_ingestion.values.config_file_path}}"
+              - --inst_schema
+              - "{{job.parameters.inst_schema}}"
           job_cluster_key: edvise-pdp-inference-pipeline-cluster
           libraries:
             - pypi:
