+++ pipelines/pdp/resources/github_pdp_training.yml
@@ -97,7 +97,9 @@ resources:
               - --job_type
               - "{{job.parameters.job_type}}"
               - --db_run_id
-              - "{{job.parameters.db_run_id}}"          
+              - "{{job.parameters.db_run_id}}"
+              - --inst_schema
+              - "{{job.parameters.inst_schema}}"
           job_cluster_key: edvise-pdp-training-pipeline-cluster
           libraries:
             - pypi:
@@ -172,6 +174,8 @@ resources:
               - "/Volumes/{{job.parameters.DB_workspace}}/{{job.parameters.databricks_institution_name}}_bronze/bronze_volume/training_inputs/{{job.parameters.config_file_name}}"
               - --db_run_id
               - "{{job.parameters.db_run_id}}"
+              - --inst_schema
+              - "{{job.parameters.inst_schema}}"
           job_cluster_key: edvise-pdp-training-pipeline-cluster
           libraries:
             - pypi:
@@ -210,6 +214,8 @@ resources:
               - "{{job.parameters.db_run_id}}"
               - --config_file_name
               - "{{job.parameters.config_file_name}}"
+              - --inst_schema
+              - "{{job.parameters.inst_schema}}"
           job_cluster_key: edvise-pdp-training-pipeline-cluster
           libraries:
             - pypi:
